refactor: 패키지 이동
- Reservation 클래스를 소스 코드로 이동 시켰습니다.
